requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0

# Data Sources
mftool>=0.2.2
yfinance>=0.2.37

# Visualization
plotly>=5.20.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
